third_party/blink/web_tests/virtual/shared_array_buffer_on_desktop/crypto/random-values-expected.txt

Tests crypto.randomValues.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'crypto' in self is true
PASS 'getRandomValues' in self.crypto is true
PASS self.crypto.__proto__.hasOwnProperty('getRandomValues') is true
PASS matchingBytes < 100 is true
PASS crypto.getRandomValues(new Uint8Array(new SharedArrayBuffer(100))) threw exception TypeError: Failed to execute 'getRandomValues' on 'Crypto': The provided ArrayBufferView value must not be shared..
PASS successfullyParsed is true

TEST COMPLETE
